backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
supabase==2.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
razorpay==1.4.2
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
slowapi==0.1.9
httpx==0.25.2
